+++ src/plotting_old/plot_efficiency.py
@@ -7,7 +7,7 @@ from tqdm import tqdm
 
 from src.env import EvacuationEnv
 from src.env import constants
-from src import params
+from src import params_old
 from main import setup_env
 from src.utils import parse_args
 from src.plotting.paths2models import *
